Address Alexey's comments
